requirements-django-extensions.txt
django-extensions==2.1.6
pygraphviz==1.5
